RGB swap finished. tested

+++ components/wmtk_components/adaptive_tessellation/wmtk/components/adaptive_tessellation/invariants/RGBSwapInvariant.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
+#include <Eigen/Dense>
 #include <wmtk/attribute/AttributeHandle.hpp>
 #include <wmtk/invariants/Invariant.hpp>
-
 namespace wmtk {
 class RGBSwapInvariant : public Invariant
 {
@@ -13,15 +13,15 @@ class RGBSwapInvariant : public Invariant
 public:
     RGBSwapInvariant(
         const Mesh& m,
-        const TypedAttributeHandle<double>& face_rgb_state_handle,
-        const TypedAttributeHandle<double>& edge_rgb_state_handle);
+        const TypedAttributeHandle<int64_t>& face_rgb_state_handle,
+        const TypedAttributeHandle<int64_t>& edge_rgb_state_handle);
     bool before(const simplex::Simplex& t) const override;
 
 private:
     bool can_swap(
         const Eigen::Vector2<int64_t>& edge_color_level,
         const Eigen::Vector2<int64_t>& face_color_level) const;
-    const TypedAttributeHandle<double> m_face_rgb_state_handle;
-    const TypedAttributeHandle<double> m_edge_rgb_state_handle;
+    const TypedAttributeHandle<int64_t> m_face_rgb_state_handle;
+    const TypedAttributeHandle<int64_t> m_edge_rgb_state_handle;
 };
 } // namespace wmtk
